backend/requirements.txt
acres==0.5.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
# Django==5.2.7
Django==5.1.4
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
etelemetry==0.3.1
filelock==3.20.0
fitz==0.0.1.dev2
httplib2==0.31.0
idna==3.11
importlib_resources==6.5.2
looseversion==1.3.0
lxml==6.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pgvector==0.4.1
pillow==11.3.0
prov==2.1.1
psycopg2-binary==2.9.11
puremagic==1.30
pycparser==2.23
pydot==4.0.1
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
rdflib==7.2.1
requests==2.32.5
scipy==1.16.2
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
tesseract==0.1.3
traits==7.0.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
langchain-text-splitters==0.3.4
sentence-transformers==3.3.1

# Transformers y PyTorch para BART/mBART
transformers==4.46.3
torch==2.5.1
sentencepiece==0.2.0
protobuf==4.25.1

# Celery y Redis para tareas asíncronas
celery==5.4.0
redis==5.2.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

# Clustering avanzado (opcional pero recomendado)
hdbscan==0.8.33
umap-learn==0.5.6

# Topic Modeling con BERTopic
bertopic==0.16.4
